+++ CHANGELOG.md
@@ -2,6 +2,15 @@
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
+## [3.0] - 2022-12-16
+### Added
+- `Loader`: general interface for `ResourceLoader`
+- `Cache`: general-purpose cache
+- `HTTPHeader`: class for parsing HTTP(s) header
+### Changed
+- `ResourceLoader`: added caching of results
+- `ResourceDescriptor`: added URL to descriptor
+
 ## [2.0] - 2022-02-07
 ### Changed
 - `ResourceLoader`: added recursion check